+++ database/migrations/2026_03_07_100936_add_employee_id_field_to_users_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('employee_id')->after('branch_ids')->nullable();
-        });
+        if (Schema::hasTable('users') && !Schema::hasColumn('users', 'employee_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->unsignedBigInteger('employee_id')->after('branch_ids')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('employee_id');
-        });
+        if (Schema::hasTable('users') && Schema::hasColumn('users', 'employee_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('employee_id');
+            });
+        }
     }
 };
